fix: do not fail on block difficulties greater than 53 bit numbers
We are running a Parity PoA network and our block difficulty is `0xfffffffffffffffffffffffffffffff6`. This difficulty is not used since it is not a PoW consensus algorithm.

When connecting a provider, ethers would error out soon after start due to `Number can only safely store up to 53 bits`. The maximum difficulty it can handle is `0x20000000000000` (https://github.com/indutny/bn.js/blob/master/lib/bn.js#L128)

This commit check the difficulty and if it goes over `0x20000000000000` it keeps it as a string and not a number.