Makes ContractInterface readonly.
This allows someone to pass in something like:
```ts
const abi = [{ ... }] as const
```
Note that when you do `as const` you get the most narrow type possible, which is immutable (everything is `readonly`).  Currently, `ethers.Contract(..., abi, ...)` will fail because it wants a mutable `abi` in, despite the fact that it doesn't actually mutate it I believe.