Added other examples
Hey there - I added in 3rd party abstractions and tools that are popular with developers to get started and building on ethers.js. I work at Alchemy and I added in a few more Provider options as well as mentioned the SDK as it's one of our developers' most used features in terms of ethers.js toolings - thought it would be helpful for your readers to refer to. Love what ethers.js libraries do for devs, pls lmk of any questions :)