Provide helpful debug on array/tuple encoding errors
If a value is missing from a tuple, currently an error will be thrown
which does not help shed any light on what went wrong, e.g. in the
case of a missing BigNumberish:

    Error: invalid BigNumber value (argument="value", value=undefined, code=INVALID_ARGUMENT, version=bignumber/5.7.0)

The developer is likely to end up confused, pondering one or more of
the following questions:

- In which part of the tuple was the problem located?

- Why did NumberCode.encode fail? Did I mistakenly pass a value which
  was supposed to be a BigNumber, or was it OK to pass a number and
  have ethers convert it?

- What does argument="value" mean?

- What does value=undefined mean?

- Did I do something else wrong?

An example of this is given in:

    https://github.com/ethers-io/ethers.js/discussions/3772

So add a try / catch to the relevant bit of abi.coders.array.pack(),
and when encoding of an individual value fails, log a helpful error
explaining which value failed, and also propagate the error message
recursively so that it gives full context of where the missing field
is situated within the overall data structure.